Create arguments mod w/ all the argument related code
